refactor: break out calendar export banner into it's own component
